+++ composeApp/src/commonMain/kotlin/dev/jketterer/leaflog/domain/models/BrewingConfiguration.kt
@@ -13,8 +13,8 @@ data class BrewingConfiguration(
 
     // Brewing parameters
     val teaQuantityGrams: Float?,
-    val waterQuantityMl: Int,
-    val temperatureCelsius: Int,
+    val waterQuantityMl: Double,
+    val temperatureCelsius: Double,
     val brewingTime: Duration,
     val waterType: WaterType,
