+++ composeApp/src/commonMain/kotlin/dev/jketterer/leaflog/presentation/ui/components/vessel/VesselCard.kt
@@ -86,7 +86,7 @@ fun VesselCard(
                     if (vessel.capacityMl != null) {
                         Text(
                             text = VolumeFormatter.format(
-                                milliliters = vessel.capacityMl,
+                                milliliters = vessel.capacityMl.toDouble(),
                                 unit = volumeUnit,
                             ),
                             style = MaterialTheme.typography.bodySmall,
