+++ composeApp/src/commonMain/kotlin/dev/jketterer/leaflog/presentation/ui/screens/settings/teatype/EditTeaTypeState.kt
@@ -32,7 +32,7 @@ data class EditTeaTypeState(
     val hasChanges: Boolean
         get() = if (isEditMode && existingTeaType != null) {
             val existingTemp = existingTeaType.defaultTemperatureCelsius?.let {
-                UnitConverter.celsiusToDisplayTemperature(it, userPreferences.temperatureUnit)
+                UnitConverter.celsiusToDisplayTemperature(it.toDouble(), userPreferences.temperatureUnit)
             }?.toString() ?: ""
             name != existingTeaType.name ||
                     colorHex != existingTeaType.colorHex ||
